ai_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
chromadb==0.6.3
pyyaml==6.0.2
pydantic==2.10.5
httpx==0.28.1
python-multipart==0.0.20
aiofiles==24.1.0
